+++ lib/events/session_event.g.dart
@@ -8,7 +8,6 @@ part of 'session_event.dart';
 
 SessionEvent _$SessionEventFromJson(Map<String, dynamic> json) {
   return SessionEvent(
-    expiresAt: json['expires_at'] as int? ?? 0,
     newKey: json['new_key'] as String?,
     ekey: json['ekey'] as String?,
     key: json['key'] as String?,