+++ lib/request/user/list/banned_user_list_request.dart
@@ -1,4 +1,5 @@
 import 'package:sendbird_sdk/constant/enums.dart';
+import 'package:sendbird_sdk/core/models/restricted_user.dart';
 import 'package:sendbird_sdk/request/abstract/api_request.dart';
 import 'package:sendbird_sdk/core/models/responses.dart';
 import 'package:sendbird_sdk/services/network/http_client.dart';
@@ -22,9 +23,12 @@ class BannedUserListRequest extends ApiRequest {
 
   @override
   Future<UserListQueryResponse> response(res) async {
-    final users =
-        List<Map>.from(res['banned_list']).map((e) => e['user']).toList();
+    final users = List<Map>.from(res['banned_list']);
+    users.forEach((e) {
+      e['type'] = 'banned';
+    });
+
     final output = {'users': users, 'next': res['next']};
-    return UserListQueryResponse.fromJson(output);
+    return UserListQueryResponse<RestrictedUser>.fromJson(output);
   }
 }